- Added assimp library to external - Fixed all projects to use windows sdk 10.0.17763.0 - Added Tools folder with project that uses assimp - Added model and meshIO - Fixed GraphicsGL to define GLEW_STATIC to fix 32 bit compile error
git-tfs-id: [https://dev.azure.com/timjustin/]$/Angazi;C261

+++ Framework/GraphicsGL/Src/MeshBufferGL.cpp
@@ -110,7 +110,7 @@ void MeshBufferGL::InitializeInternal(const void * vertices, int vertexSize, int
 	
 	auto vertexLayout = GetVertexLayout(vertexFormat);
 	UINT_PTR offset = 0;
-	for (int i = 0; i < vertexLayout.size(); i++)
+	for (size_t i = 0; i < vertexLayout.size(); i++)
 	{
 		glEnableVertexAttribArray(i);
 		glVertexAttribPointer(i, vertexLayout[i].count, vertexLayout[i].type, vertexLayout[i].normalized,
